Moved asset back
